vm -> API tests: moved misplaced runBlockchain() tests from index.spec.ts to runBlockchain.spec.ts, removed redundant test, added index.spec.ts test file description
